requirements.txt
fastapi==0.115.14
langchain==0.3.26
langchain_chroma==0.2.4
langchain_community==0.3.27
langchain_core==0.3.68
langchain_google_genai==2.1.6
langchain_huggingface==0.3.0
pandas==2.3.0
pydantic==2.11.7
python-dotenv==1.1.1
Requests==2.32.4
streamlit==1.45.1
